+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Dockerfile and distroless image.
 - Kubernetes deployment example.
 - Rules engine and rules files.