+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2021-12-04
 ### Added
 - Dockerfile and distroless image.
 - Kubernetes deployment example.
@@ -163,7 +165,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Prototype web UI.
 
-[Unreleased]: https://github.com/stefano-pogliani/auth-gateway/compare/v0.4.6...HEAD
+[Unreleased]: https://github.com/stefano-pogliani/auth-gateway/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/stefano-pogliani/auth-gateway/compare/v0.4.6...v0.5.0
 [0.4.6]: https://github.com/stefano-pogliani/auth-gateway/compare/v0.4.5...v0.4.6
 [0.4.5]: https://github.com/stefano-pogliani/auth-gateway/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/stefano-pogliani/auth-gateway/compare/v0.4.3...v0.4.4